Add new measurements from the Tibber API to the data collector
- Worker is renamed to RealtimeWorker for clarity
- A new worker is added for GraphQL queries (GraphQLWorker)

In RealtimerMeasurmentObserver:
Additional flags are added to indicate if it has had an error
or has completed. These are used by the RealtimeWorker to know
that it should resubscribe for the listener.

GrapQLWorker:
This worker collects energy price data every 5 minutes. The data is
given as prices that start at one hour intervals for today and
tommorow. Tomorrows prices do not appeare until the afternoon.
Even though this could run every hour since the data does not
change, it still goes every five minutes to reduce the wait time
when vaules change. This could be set to run at certain times of
day to reduce the amount of calls even more, but the data is so
small that it hardly matters.

Tibber Data Collector.csproj & Tibber Data Collector.xml
are related to StyleCop requirements.